requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
lunarcalendar==0.0.9
Werkzeug==3.0.1
